+++ djpcms/views/staticsite.py
@@ -45,6 +45,7 @@ def getcache(request):
     return cache
 
 
+
 def handler(request):
     '''
     Main entry for flat-pages.