+++ CHANGELOG.rst
@@ -10,4 +10,6 @@ Not yet released
  * Bug fixes
  * Added "splitregex" named options in views.appview.AppView constructor 
  * Added DISQUS in plugins
- * Removed StaticPagesMiddleware request handler and moved into views.appview.AppView.response
+ * Removed StaticPagesMiddleware request handler
+ * response method in djpcmsview class has been replaced with __call__ method
+ * Change in urls